LayoutTests/accessibility/mac/replace-text-with-empty-range-expected.txt

This tests that when we are adding text to empty text fields, the accessibility replace with range API functions as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axContentEditableDiv.replaceTextInRange('Hello', 0, 0) is true
PASS axContentEditableDiv.stringValue is 'AXValue: Hello'
PASS axText.replaceTextInRange('Hello', 0, 0) is true
PASS axText.stringValue is 'AXValue: Hello'
PASS axTextarea.replaceTextInRange('Hello', 0, 0) is true
PASS axTextarea.stringValue is 'AXValue: Hello'
PASS successfullyParsed is true

TEST COMPLETE
